Release

+++ src/main/java/kr/codeback/service/impl/CodeReviewCommentServiceImpl.java
@@ -5,23 +5,26 @@ import java.time.LocalDate;
 import java.util.List;
 import java.util.UUID;
 
-import kr.codeback.model.dto.request.review.CodeReviewCommentRequestDTO;
-import kr.codeback.model.dto.request.review.CommentModifyRequestDTO;
-import kr.codeback.model.dto.response.review.CodeReviewCommentResponseDTO;
-import kr.codeback.model.entity.*;
-import kr.codeback.repository.CodeReviewRepository;
-import kr.codeback.repository.MemberRepository;
 import org.springframework.stereotype.Service;
 
 import jakarta.transaction.Transactional;
-import kr.codeback.model.dto.response.summary.CodeReviewCommentSummaryResponseDTO;
+import kr.codeback.exception.ErrorCode;
+import kr.codeback.exception.member.MemberNotFoundException;
+import kr.codeback.exception.review.ReviewNonExistentException;
+import kr.codeback.exception.review.ReviewNotAuthorizedException;
+import kr.codeback.model.dto.request.review.CodeReviewCommentRequestDTO;
+import kr.codeback.model.dto.request.review.CommentModifyRequestDTO;
+import kr.codeback.model.dto.response.summary.SummaryByMonthResponseDTO;
 import kr.codeback.model.entity.CodeReview;
 import kr.codeback.model.entity.CodeReviewComment;
 import kr.codeback.model.entity.Member;
+import kr.codeback.model.entity.Preference;
 import kr.codeback.repository.CodeReviewCommentRepository;
+import kr.codeback.repository.CodeReviewRepository;
+import kr.codeback.repository.MemberRepository;
 import kr.codeback.service.interfaces.CodeReviewCommentService;
-import kr.codeback.service.interfaces.CodeReviewPreferenceService;
 import kr.codeback.service.interfaces.NotificationService;
+import kr.codeback.service.interfaces.PreferenceService;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -30,11 +33,10 @@ public class CodeReviewCommentServiceImpl implements CodeReviewCommentService {
 
 	private final CodeReviewCommentRepository codeReviewCommentRepository;
 
-	private final CodeReviewPreferenceService codeReviewPreferenceService;
+	private final PreferenceService preferenceService;
 	private final NotificationService notificationService;
 	private final MemberRepository memberRepository;
 	private final CodeReviewRepository codeReviewRepository;
-
 
 	@Override
 	@Transactional
@@ -48,11 +50,11 @@ public class CodeReviewCommentServiceImpl implements CodeReviewCommentService {
 
 		codeReviewComments.stream()
 			.map(CodeReviewComment::getId)
-			.forEach(notificationService::deleteByEntityID);
+			.forEach(notificationService::deleteByEntityId);
 
 		codeReviewComments.stream()
 			.map(CodeReviewComment::getId)
-			.forEach(codeReviewPreferenceService::deleteByEntityID);
+			.forEach(preferenceService::deleteByEntityID);
 
 		codeReviewCommentRepository.deleteAll(codeReviewComments);
 	}
@@ -65,45 +67,59 @@ public class CodeReviewCommentServiceImpl implements CodeReviewCommentService {
 
 		codeReviewComments.stream()
 			.map(CodeReviewComment::getId)
-			.forEach(notificationService::deleteByEntityID);
+			.forEach(notificationService::deleteByEntityId);
 
 		codeReviewComments.stream()
 			.map(CodeReviewComment::getId)
-			.forEach(codeReviewPreferenceService::deleteByEntityID);
+			.forEach(preferenceService::deleteByEntityID);
 
 		codeReviewCommentRepository.deleteAll(codeReviewComments);
 	}
 
 	@Override
-	public CodeReviewCommentResponseDTO saveComment(CodeReviewCommentRequestDTO commentDTO) {
+	public CodeReviewComment saveComment(CodeReviewCommentRequestDTO commentDTO) {
 		Member member = memberRepository.findByEmail(commentDTO.getMemberEmail())
-				.orElseThrow(()-> new IllegalArgumentException("no member..."));
-		CodeReview codeReview = codeReviewRepository.findById(commentDTO.getCodeReviewId())
-				.orElseThrow(()-> new IllegalArgumentException("no codeReview..."));
+			.orElseThrow(() -> new MemberNotFoundException(
+				ErrorCode.NOT_EXIST_USER.getStatus(),
+				ErrorCode.NOT_EXIST_USER.getMessage()
+			));
+
+		CodeReview codeReview = codeReviewRepository.findById(commentDTO.getReviewId())
+			.orElseThrow(() -> new ReviewNonExistentException(
+				ErrorCode.NONEXISTENT_REVIEW.getStatus(),
+				ErrorCode.NOT_EXIST_USER.getMessage()
+			));
 
 		CodeReviewComment codeReviewComment = CodeReviewComment.builder()
-				.comment(commentDTO.getContent())
-				.id(UUID.randomUUID())
-				.member(member)
-				.codeReview(codeReview)
-				.build();
+			.comment(commentDTO.getContent())
+			.id(UUID.randomUUID())
+			.member(member)
+			.codeReview(codeReview)
+			.build();
 		codeReviewCommentRepository.save(codeReviewComment);
 
-		CodeReviewComment savedComment = codeReviewCommentRepository.findById(codeReviewComment.getId())
-				.orElseThrow(()->new IllegalArgumentException("no comment..."+codeReviewComment.getId()));
-
-		return savedComment.toDTO();
+		return codeReviewComment;
 	}
 
 	@Override
-	public void deleteById(UUID commentId) {
+	public void deleteById(UUID commentId, String memberEmail) {
 		CodeReviewComment comment = codeReviewCommentRepository.findById(commentId)
-				.orElseThrow(()->new IllegalArgumentException("no comments.."+commentId));
-		List<CodeReviewPreference> preferences = codeReviewPreferenceService.findByEntityID(commentId);
-		codeReviewPreferenceService.deleteAll(preferences);
+			.orElseThrow(() -> new ReviewNonExistentException(
+				ErrorCode.NONEXISTENT_REVIEW.getStatus(),
+				ErrorCode.NOT_EXIST_USER.getMessage()
+			));
 
-		List<Notification> notifications = notificationService.findByEntityID(commentId);
-		notificationService.deleteAll(notifications);
+		if (!comment.getMember().getEmail().equals(memberEmail)) {
+			throw new ReviewNotAuthorizedException(
+				ErrorCode.NOT_EXIST_USER.getStatus(),
+				ErrorCode.NOT_EXIST_USER.getMessage()
+			);
+		}
+
+		List<Preference> preferences = preferenceService.findByEntityID(commentId);
+		preferenceService.deleteAll(preferences);
+
+		notificationService.deleteByEntityId(commentId);
 
 		codeReviewCommentRepository.delete(comment);
 	}
@@ -111,15 +127,26 @@ public class CodeReviewCommentServiceImpl implements CodeReviewCommentService {
 	@Override
 	public void update(CommentModifyRequestDTO commentDTO) {
 		CodeReviewComment comment = codeReviewCommentRepository.findById(commentDTO.getId())
-				.orElseThrow(()->new IllegalArgumentException("no Comment..."));
+			.orElseThrow(() -> new ReviewNonExistentException(
+				ErrorCode.NOT_EXIST_USER.getStatus(),
+				ErrorCode.NOT_EXIST_USER.getMessage()
+			));
+
+		if (!commentDTO.getMemberEmail().equals(comment.getMember().getEmail())) {
+			throw new ReviewNotAuthorizedException(
+				ErrorCode.NOT_EXIST_USER.getStatus(),
+				ErrorCode.NOT_EXIST_USER.getMessage()
+			);
+		}
 
 		comment.updateCodeReviewComment(commentDTO);
 		codeReviewCommentRepository.save(comment);
-  }
+	}
 
-  @Override
-	public List<CodeReviewCommentSummaryResponseDTO> calculateSummaryByMonth(String inputDate) {
-		Date searchDate = null;
+	@Override
+	public List<SummaryByMonthResponseDTO> calculateSummaryByMonth(String inputDate) {
+
+		Date searchDate;
 		if (inputDate == null || inputDate.isEmpty()) {
 			searchDate = Date.valueOf(LocalDate.now());
 		} else {
@@ -129,7 +156,7 @@ public class CodeReviewCommentServiceImpl implements CodeReviewCommentService {
 		List<Object[]> results = codeReviewCommentRepository.calculateSummaryByMonth(searchDate);
 
 		return results.stream().map(
-				row -> new CodeReviewCommentSummaryResponseDTO(
+				row -> new SummaryByMonthResponseDTO(
 					Integer.parseInt(row[0].toString()),
 					((Number)row[1]).longValue()
 				))
